Update imports and improve readability of the solution form
Updated import paths in `AdminProblemList.vue`, `ProblemForm.vue`, and `AdminDashboard.vue` to reflect the new directory structure. Improved the organization and readability of the SolutionForm component.
These changes make the code more maintainable and easier to understand.
