+++ .gitignore
@@ -1,4 +1,5 @@
 .idea/
 /node_modules/
+/logs/
 .env
 vite.config.ts.timestamp-*
